Fixed a bunch of compiler warnings.

+++ src/lib/config/A_Command_Line_Parser.cpp
@@ -33,7 +33,7 @@ bool A_Command_Line_Parser::Check_If_Flag_Exists( const std::string& flag )
 {
 
     // Iterate through flags
-    for( int i=0; i<m_argument_list.size(); i++ ){
+    for( int i=0; i<(int)m_argument_list.size(); i++ ){
         if( flag == m_argument_list[i] ){
             m_reference_list[i]++;
             return true;
@@ -55,14 +55,14 @@ std::string A_Command_Line_Parser::Query_Flag_Value_As_String( const std::string
     value_found = false;
 
     // Iterate through flags
-    for( int i=0; i<m_argument_list.size(); i++ )
+    for( int i=0; i<(int)m_argument_list.size(); i++ )
     {
         // Check if flag found
         if( flag == m_argument_list[i] ){
             
             flag_found = true;
 
-            if( i < (m_argument_list.size()-1) && m_reference_list[i+1] < 1 ){
+            if( i < (((int)m_argument_list.size())-1) && m_reference_list[i+1] < 1 ){
                 value_found = true;
                 m_reference_list[i]++;
                 return m_argument_list[i+1];
